feat(WW-3531): integrate submitForm from useForm for auto-submit functionality
